+++ src/common/components/input/input.module.scss
@@ -87,6 +87,7 @@
   display: block;
   width: 100%;
   min-width: 284px;
+  max-width: 100%;
   height: 36px;
 
   // Disable clear button for input with type='search'
