+++ src/features/auth/ui/newPasswordForm/newPasswordForm.module.scss
@@ -1,5 +1,5 @@
 .card {
-  padding: 35px 35px 60px 35px ;
+  padding: 35px 35px 60px 35px;
   width: 100%;
   max-width: 420px;
 }
