+++ src/features/auth/ui/signUp/formSignUp.module.scss
@@ -1,5 +1,5 @@
 .card {
-  padding: 35px 35px 25px 35px;
+  padding: 35px 35px 25px;
   max-width: 420px;
   width: 100%;
 }
