--- src/features/modals/modalAddEditForm/controlledInputFile.module.scss
@@ -1,8 +1,0 @@
-.file {
-  display: none;
-}
-
-.label {
-  text-align: center;
-  width: 100%;
-}
