+++ src/features/modals/ui/controlledInputFile.module.scss
@@ -1,0 +1,28 @@
+.file {
+  display: none;
+}
+
+.label {
+  text-align: center;
+  width: 100%;
+}
+
+.img {
+  object-fit: cover;
+  object-position: 50% 50%;
+  max-height: 160px;
+  max-width: 480px;
+  width: 100%;
+  height: 100%;
+}
+
+.container {
+  display: grid;
+  flex-direction: column;
+  gap: 24px;
+}
+.containerButton {
+  display: flex;
+  justify-content: space-between;
+  gap: 24px;
+}
